refactor(gateway): restore rust standards baseline

+++ code/apps/bpane-gateway/src/transport/egress.rs
@@ -64,7 +64,7 @@ pub(super) fn spawn_agent_to_browser_task(
                     }
                 }
                 Err(broadcast::error::RecvError::Lagged(n)) => {
-                    ctx.hub.record_egress_lagged(n as u64);
+                    ctx.hub.record_egress_lagged(n);
                     warn!(
                         ctx.session_id,
                         ctx.client_id, n, "client lagged, skipping frames"
